cuboid: Change rounding algorithm for "else" case
For some reason, the old code would be *really* slow in preview compared to the
minkowski shape when edges=EDGES_ALL.

This changes it to intersect with a minkowski'd cuboid that is sized so
that the rounded edges don't intersect with edges that aren't specified in
`edges`.

The result is that the preview goes a lot faster, but unfortunately doesn't
cover the same use cases as before -- it is unable to handle 2 corner roundings,
e.g.

    cuboid(rounding=0.5, edges=RIGHT)

will now look like

    cuboid(rounding=0.5, edges=edges("ALL", except=LEFT)

But that is acceptable for my usage for now. Maybe there should be a check to
see which implementation it uses depending on what edges are selected (whole
faces or not).